+++ src/apsg/__init__.py
@@ -162,6 +162,6 @@ __all__ = (
     "quicknet",
 )
 
-__version__ = "1.4.0"
+__version__ = "1.4.1"
 __author__ = "Ondrej Lexa"
 __email__ = "lexa.ondrej@gmail.com"